+++ backend/main.go
@@ -1,46 +1,27 @@
 package main
 
 import (
-	"log"
-	"net/http"
+	"latihan-assessment/backend/helper"
 
 	"github.com/gin-gonic/gin"
-	"gorm.io/gorm"
-
-	"latihan-assessment/backend/config"
 )
-
-var DB *gorm.DB = config.Conn()
 
 func main() {
 	r := gin.Default()
 
-	r.GET("/users", HandleUsers)
+	r.GET("/users", helper.HandleUsers)
+	r.POST("/users/", helper.HandleUsersPOST)
 	r.POST("/users/register")
 	r.POST("/users/login")
-	r.GET("/users/:id")
+	r.GET("/users/:id", helper.HandleUserID)
 	r.PUT("/users/:id")
 	r.DELETE("/users/:id")
 
-	r.GET("/books", HandleBook)
+	r.GET("/books", helper.HandleBook)
 	r.POST("/books/")
 	r.GET("/books/:id")
 	r.PUT("/books/:id")
 	r.DELETE("/books/:id")
 
 	r.Run(":8080")
-}
-
-func HandleUsers(c *gin.Context) {
-	var users []Users
-
-	if err := DB.Find(&users).Error; err != nil {
-		log.Println(err.Error())
-	}
-
-	c.JSON(http.StatusOK, users)
-}
-
-func HandleBook(c *gin.Context) {
-
 }
